add weapon aquire system
이제 무기는 획득을 해야만 사용할 수 있다.
(Script_GroundPlayer::AquireWeapon)

+++ Client/X-MACHINA/src/Script_GameManager.cpp
@@ -50,7 +50,7 @@ void Script_GameManager::Awake()
 			auto& script = enemy->AddComponent<Script_Onyscidus>();
 			enemy->AddComponent<Script_MeleeBT>();
 			enemy->SetPosition(pos);
-			enemy->Rotate(0, rand() % 360, 0);
+			enemy->Rotate(0, (float)(rand() % 360), 0);
 		}
 	}
 
@@ -70,7 +70,7 @@ void Script_GameManager::Awake()
 			auto& script = enemy->AddComponent<Script_AdvancedCombatDroid_5>();
 			enemy->AddComponent<Script_MeleeBT>();
 			enemy->SetPosition(pos);
-			enemy->Rotate(0, rand() % 360, 0);
+			enemy->Rotate(0, (float)(rand() % 360), 0);
 		}
 	}
 }
